oauth: add rpc scope for service-auth minting; share the scope list
getServiceAuth returns 403 under the granular OAuth model unless the token
holds a matching `rpc` grant. The console mints service-auth JWTs (from the
user's session) for dev.cocore.devicePair.confirm and dev.cocore.inference
.dispatch so the AppView can verify the caller's DID — but the scope list
only granted account/blob/repo, so those mints 403'd. This is why device
pairing failed at "getServiceAuth returned 403".

- Add `rpc?lxm=devicePair.confirm,inference.dispatch&aud=*` to the scope.
- Collapse the two hand-mirrored scope copies (console minting client +
  AppView restoring client) into one shared `@cocore/sdk/oauth-scope`
  constant; both import it so they can't drift.

NOTE: existing users must RE-AUTHORIZE (sign out + sign in) to get the new
rpc grant — tokens issued before this don't retroactively gain it.
